Check results file for failures in Makefile flow
