+++ docs/source/newsfragments/4337.change.rst
@@ -1,0 +1,1 @@
+The Makefile test flow now fails the ``sim`` target if there are test failures.
