+++ src/cocotb_tools/combine_results.py
@@ -1,3 +1,6 @@
+# Copyright cocotb contributors
+# Licensed under the Revised BSD License, see LICENSE for details.
+# SPDX-License-Identifier: BSD-3-Clause
 #!/usr/bin/env python
 """
 Simple script to combine JUnit test results into a single XML file.
