+++ src/cocotb_tools/makefiles/Makefile.inc
@@ -72,6 +72,11 @@ define check_for_results_file
     @test -f $(COCOTB_RESULTS_FILE) || (echo "ERROR: $(COCOTB_RESULTS_FILE) was not written by the simulation!" >&2 && exit 1)
 endef
 
+# Check that the COCOTB_RESULTS_FILE was created and has no failures
+define check_results
+    @$(PYTHON_BIN) -m cocotb_tools.check_results $(COCOTB_RESULTS_FILE)
+endef
+
 SIM_BUILD ?= sim_build
 export SIM_BUILD
