+++ src/cocotb_tools/makefiles/simulators/Makefile.questa
@@ -165,7 +165,7 @@ $(COCOTB_RESULTS_FILE): $(SIM_BUILD)/runsim.do
 	GPI_EXTRA=$(GPI_EXTRA) TOPLEVEL_LANG=$(TOPLEVEL_LANG) \
 	$(SIM_CMD_PREFIX) $(CMD) $(RUN_ARGS) -do $(SIM_BUILD)/runsim.do $(call deprecate,PLUSARGS,COCOTB_PLUSARGS) $(SIM_CMD_SUFFIX)
 
-	$(call check_for_results_file)
+	$(call check_results)
 
 clean::
 	$(RM) -r $(SIM_BUILD)
