+++ src/cocotb_tools/makefiles/simulators/Makefile.xcelium
@@ -127,7 +127,7 @@ $(COCOTB_RESULTS_FILE): $(HDL_SOURCES) $(CUSTOM_COMPILE_DEPS) $(CUSTOM_SIM_DEPS)
 	-vhdl_time_precision "$(COCOTB_HDL_TIMEPRECISION)" \
 	$(EXTRA_ARGS) $(GPI_ARGS) $(INCDIRS) -access +rwc -createdebugdb $(LIBS) $(MAKE_LIB) $(HDL_SOURCES) $(END_LIB) $(call deprecate,PLUSARGS,COCOTB_PLUSARGS) $(SIM_CMD_SUFFIX)
 
-	$(call check_for_results_file)
+	$(call check_results)
 
 clean::
 	$(RM) -r $(SIM_BUILD)
